ChatModels/chatmodel_hf_api.py
from langchain_huggingface import ChatHuggingFace , HuggingFaceEndpoint
from dotenv import load_dotenv
load_dotenv()


llm = HuggingFaceEndpoint(
    repo_id="google/gemma-2-2b-it",
    task="text-generation",
)

model = ChatHuggingFace(llm=llm)

result = model.invoke("give me 5 line info on indian prime minister Narendra Modi")

print(result.content)