requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Hugging face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment variable Management
python-dotenv

# machine Learning Institute
numpy
scikit-learn